requirements.txt
httpx[http2]>=0.24
fastapi>=0.110
uvicorn[standard]>=0.22
pydantic>=1.10
pydantic-settings>=2.0
python-dotenv>=1.0
gunicorn>=21.0
openpyxl>=3.1.5
python-multipart>=0.0.9
pymupdf>=1.24.0
python-docx>=1.1.0
numpy>=1.24.0
